+++ sdk/containerregistry/mgmt/src/main/java/com/azure/management/containerregistry/Registry.java
@@ -12,7 +12,7 @@ import com.azure.management.resources.fluentcore.model.Appliable;
 import com.azure.management.resources.fluentcore.model.Creatable;
 import com.azure.management.resources.fluentcore.model.Refreshable;
 import com.azure.management.resources.fluentcore.model.Updatable;
-import com.azure.management.storage.StorageAccount;
+import com.azure.management.storage.models.StorageAccount;
 import java.time.OffsetDateTime;
 import java.util.Collection;
 import reactor.core.publisher.Mono;