+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/rx/UniqueIndexTest.java
@@ -19,10 +19,8 @@ import com.azure.cosmos.implementation.guava25.collect.Lists;
 import com.azure.cosmos.implementation.models.CosmosItemResponseImpl;
 import com.azure.cosmos.models.CosmosContainerProperties;
 import com.azure.cosmos.models.CosmosItemRequestOptions;
-import com.azure.cosmos.models.DataType;
 import com.azure.cosmos.models.ExcludedPath;
 import com.azure.cosmos.models.IncludedPath;
-import com.azure.cosmos.models.Index;
 import com.azure.cosmos.models.IndexingMode;
 import com.azure.cosmos.models.IndexingPolicy;
 import com.azure.cosmos.models.ModelBridgeInternal;
@@ -75,11 +73,8 @@ public class UniqueIndexTest extends TestSuiteBase {
         indexingPolicy.setExcludedPaths(Collections.singletonList(excludedPath));
 
         IncludedPath includedPath1 = new IncludedPath("/name/?");
-        includedPath1.setIndexes(Collections.singletonList(Index.hash(DataType.STRING, 7)));
-        includedPath1.setIndexes(Collections.singletonList(Index.hash(DataType.STRING, 7)));
 
         IncludedPath includedPath2 = new IncludedPath("/description/?");
-        includedPath2.setIndexes(Collections.singletonList(Index.hash(DataType.STRING, 7)));
         indexingPolicy.setIncludedPaths(ImmutableList.of(includedPath1, includedPath2));
         collectionDefinition.setIndexingPolicy(indexingPolicy);
 
@@ -187,10 +182,8 @@ public class UniqueIndexTest extends TestSuiteBase {
         indexingPolicy.setExcludedPaths(Collections.singletonList(excludedPath));
 
         IncludedPath includedPath1 = new IncludedPath("/name/?");
-        includedPath1.setIndexes(Collections.singletonList(Index.hash(DataType.STRING, 7)));
 
         IncludedPath includedPath2 = new IncludedPath("/description/?");
-        includedPath2.setIndexes(Collections.singletonList(Index.hash(DataType.STRING, 7)));
         indexingPolicy.setIncludedPaths(ImmutableList.of(includedPath1, includedPath2));
 
         collectionDefinition.setIndexingPolicy(indexingPolicy);