+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/implementation/converters/RequestOptionsConverter.java
@@ -22,7 +22,7 @@ public final class RequestOptionsConverter {
         RequestOptions requestOptions = new RequestOptions();
 
         UUID xMsClientRequestId = obj.getXMsClientRequestId();
-        requestOptions.setXMsClientRequestId(xMsClientRequestId);
+        requestOptions.setClientRequestId(xMsClientRequestId);
         return requestOptions;
     }
 
@@ -36,7 +36,7 @@ public final class RequestOptionsConverter {
         com.azure.search.documents.implementation.models.RequestOptions requestOptions =
             new com.azure.search.documents.implementation.models.RequestOptions();
 
-        UUID xMsClientRequestId = obj.getXMsClientRequestId();
+        UUID xMsClientRequestId = obj.getClientRequestId();
         requestOptions.setXMsClientRequestId(xMsClientRequestId);
         return requestOptions;
     }