+++ sdk/storage/mgmt/src/main/java/com/azure/management/storage/implementation/ManagementPoliciesImpl.java
@@ -4,17 +4,18 @@
 package com.azure.management.storage.implementation;
 
 import com.azure.management.resources.fluentcore.model.implementation.WrapperImpl;
-import com.azure.management.storage.ManagementPolicies;
-import com.azure.management.storage.ManagementPolicy;
-import com.azure.management.storage.models.ManagementPoliciesInner;
-import com.azure.management.storage.models.ManagementPolicyInner;
+import com.azure.management.storage.StorageManager;
+import com.azure.management.storage.fluent.ManagementPoliciesClient;
+import com.azure.management.storage.models.ManagementPolicies;
+import com.azure.management.storage.models.ManagementPolicy;
+import com.azure.management.storage.fluent.inner.ManagementPolicyInner;
 import reactor.core.publisher.Mono;
 
-class ManagementPoliciesImpl extends WrapperImpl<ManagementPoliciesInner> implements ManagementPolicies {
+public class ManagementPoliciesImpl extends WrapperImpl<ManagementPoliciesClient> implements ManagementPolicies {
     private final StorageManager manager;
 
-    ManagementPoliciesImpl(StorageManager manager) {
-        super(manager.inner().managementPolicies());
+    public ManagementPoliciesImpl(StorageManager manager) {
+        super(manager.inner().getManagementPolicies());
         this.manager = manager;
     }