+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Release History
 
 ## 1.0.0-beta.3 (Unreleased)
+- Rename parameters data and sourceUrl parameters found on methods for FormRecognizerClient to form and formUrl, respectively.
+- Rename parameters for receipt API methods to receipt and receiptUrl.
+- Raise `HttpResponseException` when a model with `ModelStatus.Invalid` is returned from the `beginTraining()` API's
+- Fix `HttpResponseException` to include the error object thrown on invalid analyze status for recognize API's
+- Update FormField property `transactionTime` on `USReceipt` to return `LocalTime` instead of `String`
 - Rename model `PageRange` to `FormPageRange`
 - Rename property `startPageNumber` to `firstPageNumber` and `endPageNumber` to `lastPageNumber` in model `PageRange`
 - Rename `getCustomModelInfos` to `listCustomModels`
@@ -19,6 +24,9 @@
 - Rename parameter `useLabelFile` to `useTrainingLabels` on `beginTraining` method in FormTrainingClients
 - Replace parameters `filePrefix` and `includeSubFolders` with `TrainingFileFilter` model
 - Rename AccountProperties `count` and `limit` to `customModelCount` and `customModelLimit`
+
+### New Features
+- Support to copy a custom model from one Form Recognizer resource to another
 
 ## 1.0.0-beta.2 (2020-05-06)
 - Fixed Receipt type bug to select the valueString field via fieldValue.